+++ events-manager-import-export.php
@@ -33,8 +33,14 @@ if (!defined('ABSPATH')) {
 define('EM_IMPEXP_PLUGIN_FILE', __FILE__);
 define('EM_IMPEXP_PLUGIN_ROOT', dirname(__FILE__) . '/');
 define('EM_IMPEXP_PLUGIN_NAME', basename(dirname(__FILE__)) . '/' . basename(__FILE__));
+define('EM_IMPEXP_MIN_PHP', '5.6');
 define('EM_IMPEXP_PLUGIN_VERSION', '0.0.12');
 
-// instantiate the plug-in
-require EM_IMPEXP_PLUGIN_ROOT . 'includes/class.EM_ImpExpPlugin.php';
-EM_ImpExpPlugin::getInstance();
+require EM_IMPEXP_PLUGIN_ROOT . 'includes/functions-global.php';
+
+if (version_compare(PHP_VERSION, EM_IMPEXP_MIN_PHP, '<')) {
+	add_action('admin_notices', 'em_impexp_fail_php_version');
+	return;
+}
+
+require EM_IMPEXP_PLUGIN_ROOT . 'includes/bootstrap.php';