+++ changelog.md
@@ -4,6 +4,8 @@
 
 ### 0.0.12, soon...
 
+* changed: requires PHP 5.6+ (recommend PHP 7.2+)
+* changed: parsecsv upgraded to v1.0.0
 * merged: [@asmartin pull request](https://github.com/webaware/events-manager-import-export/pull/1); thanks!
   - post_id is now required in the database, so add it
   - verify new location matches existing by comparing postcodes